// src/app/models/regions.model.ts
export interface Regions {
  adminCode1: string;
  adminName1: string;
  countryCode: string;
  countryId: string;
  countryName: string;
  fcl: string;
  fclName: string;
  fcode: string;
  fcodeName: string;
  geonameId: number;
  lat: string;
  lng: string;
  name: string;
  numberOfChildren: number;
  population: number;
  toponymName: string;
}