// src/app/models/country_data.model.ts
export interface CountryData {
  Continent: string;
  CurrencyCode: string;
  CurrencyName: string;
  EquivalentFipsCode: string;
  ISO: string;
  ISO3: string;
  ISO_Numeric: string;
  Languages: string;
  Phone: string;
  Population: string;
  Postal_Code_Format: string;
  Postal_Code_Regex: string;
  area: string;
  capital: string;
  country: string;
  fips: string;
  geonameid: string;
  neighbours: string;
  tld: string;
}